refactor: action module
